Modify Markdown to R-markdown

generate rmarkdown file in output folder

Usage

generate_rmd(article_dir)

Arguments

article_dir

path to the directory which contains tex article

Value

R-markdown file in the web folder

Examples

# Note This is a minimal example to execute this function
article_dir <- system.file("examples/article",
                 package = "texor")
dir.create(your_article_folder <- file.path(tempdir(), "tempdir2"))
x <- file.copy(from = article_dir, to = your_article_folder,recursive = TRUE,)
your_article_path <-  xfun::normalize_path(paste(your_article_folder,"article",sep="/"))
texor::include_style_file(your_article_path)
#> [1] TRUE
rebib::aggregate_bibliography(your_article_path)
#> Warning: No Bib files found !
#> Cannot aggregate bibliography as there is no BibTeX bibliography
#> Warning: No Bib files found !
#> BibTeX file does not exist
#> will parse for bibliography
#> bibtex file created
data <- texor:::handle_figures(your_article_path,
                    texor::get_texfile_name(your_article_path))
#> Found 0 Matches for target :  figure\*
#> Warning: Changed \begin{figure\*} to \begin{figure}
#> Found 0 Matches for target :  figure\*
#> Warning: Changed \end{figure\*} to \end{figure}
#> Found 0 Matches for target :  algorithmic
#> Warning: Changed \begin{algorithmic} to \begin{algorithm}
#> Found 0 Matches for target :  algorithmic
#> Warning: Changed \end{algorithmic} to \end{algorithm}
#> /usr/local/bin/pandoc +RTS -K512m -RTS /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpNsC5G4/tempdir2/article/RJwrapper.tex --to native --from latex --output /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpNsC5G4/tempdir2/article/temp-native.txt -s --resource-path /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpNsC5G4/tempdir2/article/RJwrapper.tex --lua-filter /Users/runner/work/_temp/Library/texor/find_pdf_files.lua 
texor:::patch_code_env(your_article_path) # Step 4
#> Found  1 Matches for target : example
#> Found  1 Matches for target : example
#> Found 0 Matches for target :  example\*
#> Found 0 Matches for target :  example\*
#> Found 0 Matches for target :  Sin
#> Found 0 Matches for target :  Sin
#> Found 0 Matches for target :  Sout
#> Found 0 Matches for target :  Sout
#> Found 0 Matches for target :  Scode
#> Found 0 Matches for target :  Scode
#> Found 0 Matches for target :  Sinput
#> Found 0 Matches for target :  Sinput
#> Found 0 Matches for target :  Soutput
#> Found 0 Matches for target :  Soutput
#> Found 0 Matches for target :  smallverbatim
#> Found 0 Matches for target :  smallverbatim
#> Found 0 Matches for target :  boxedverbatim
#> Found 0 Matches for target :  boxedverbatim
texor:::patch_table_env(your_article_path) # Step 5
#> Found  1 Matches for target : table\*
#> Warning: Changed \begin{table\*} to \begin{table}
#> Found  1 Matches for target : table\*
#> Warning: Changed \end{table*} to \end{table}
#> Found  2 Matches for target : multicolumn
#> Warning: changed \multicolumn to \multicolumnx
texor:::patch_equations(your_article_path) # Step 5.5
texor:::patch_figure_env(your_article_path)
#> Found 0 Matches for target :  figure\*
#> Warning: Changed \begin{figure\*} to \begin{figure}
#> Found 0 Matches for target :  figure\*
#> Warning: Changed \end{figure\*} to \end{figure}
#> Found 0 Matches for target :  algorithmic
#> Warning: Changed \begin{algorithmic} to \begin{algorithm}
#> Found 0 Matches for target :  algorithmic
#> Warning: Changed \end{algorithmic} to \end{algorithm}
#> Found 0 Matches for target :  algorithm
#> Warning: Changed \begin{algorithm} to \begin{figure}
#> Found 0 Matches for target :  algorithm
#> Warning: Changed \end{algorithm} to \end{figure}
rmarkdown::pandoc_version()
#> [1] ‘2.19.2’
list.files(your_article_path)
#>  [1] "DESCRIPTION"                     "Metafix.sty"                    
#>  [3] "RJwrapper.tex"                   "RJwrapper.tex.bk"               
#>  [5] "Rlogo-5.png"                     "example.bib"                    
#>  [7] "example.tex"                     "example.tex.bk"                 
#>  [9] "expected-texor-figure-meta.yaml" "temp-native.txt"                
#> [11] "texor-figure-meta.yaml"         
texor::convert_to_markdown(your_article_path)
#> /usr/local/bin/pandoc +RTS -K512m -RTS /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpNsC5G4/tempdir2/article/RJwrapper.tex --to markdown-simple_tables-pipe_tables-fenced_code_attributes --from latex --output /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpNsC5G4/tempdir2/article/RJwrapper.md -s --resource-path /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpNsC5G4/tempdir2/article --lua-filter /Users/runner/work/_temp/Library/texor/abs_filter.lua --lua-filter /Users/runner/work/_temp/Library/texor/bib_filter.lua --lua-filter /Users/runner/work/_temp/Library/texor/image_filter.lua --lua-filter /Users/runner/work/_temp/Library/texor/R_code.lua --lua-filter /Users/runner/work/_temp/Library/texor/image_caption.lua --lua-filter /Users/runner/work/_temp/Library/texor/table_caption.lua --lua-filter /Users/runner/work/_temp/Library/texor/conversion_compat_check.lua --lua-filter /Users/runner/work/_temp/Library/texor/equation_filter.lua --lua-filter /Users/runner/work/_temp/Library/texor/bookdown_ref.lua --citeproc 
#> Detected the following packages from article:
#>   CRAN: 
#>   Bioconductor: 
list.files(your_article_path)
#>  [1] "DESCRIPTION"                     "Metafix.sty"                    
#>  [3] "RJwrapper.md"                    "RJwrapper.tex"                  
#>  [5] "RJwrapper.tex.bk"                "Rlogo-5.png"                    
#>  [7] "example.bib"                     "example.tex"                    
#>  [9] "example.tex.bk"                  "expected-texor-figure-meta.yaml"
#> [11] "pkg_meta.yaml"                   "post-conversion-meta.yaml"      
#> [13] "temp-native.txt"                 "texor-figure-meta.yaml"         
cat(readLines(paste(your_article_path,"/RJwrapper.md",sep="")),sep="\n")
#> ---
#> address:
#> - |
#>   Author One\
#>   Affiliation\
#>   Address\
#>   Country\
#>   (ORCiD if desired)\
#> - |
#>   Author Two\
#>   Affiliation\
#>   Address\
#>   Country\
#>   (ORCiD if desired)\
#> - |
#>   Author Three\
#>   Affiliation\
#>   Address\
#>   Country\
#>   (ORCiD if desired)\
#> author:
#> - by Author One, Author Two and Author Three
#> - An abstract of less than 150 words.
#> bibliography:
#> - example.bib
#> title: Capitalized Title Here
#> ---
#> 
#> ::: article
#> # Section title in sentence case
#> 
#> Introductory section which may include references in parentheses [@R],
#> or cite a reference such as @R in the text.
#> 
#> # Another section
#> 
#> This section may contain a figure such as Figure [1](#figure:rlogo).
#> 
#> ![Figure 1 The logo of R.](Rlogo-5.png){#figure:rlogo .center
#> width="100%"}
#> 
#> # Table example section
#> 
#> ::: center
#> ::: {#table1}
#>   -------------------------------------
#>   EXAMPLE   \(X\)         \(Y\)   
#>   --------- ------- ----- ------- -----
#>             1       2     1       2
#> 
#>   EX1       X11     X12   Y11     Y12
#> 
#>   EX2       X21     X22   Y21     Y22
#> 
#>   EX3       X31     X32   Y31     Y32
#> 
#>   EX4       X41     X42   Y41     Y42
#> 
#>   EX5       X51     X52   Y51     Y52
#>   -------------------------------------
#> 
#>   : Table 1: An Example Table with Multicolumn
#> :::
#> 
#> []{#table1 label="table1"}
#> :::
#> 
#> # Another section
#> 
#> There will likely be several sections, perhaps including code snippets,
#> such as:
#> 
#> ``` r
#>   x <- 1:10
#>   result <- myFunction(x)
#> ```
#> 
#> # Summary
#> 
#> This file is only a basic article template. For full details of *The R
#> Journal* style and information on how to prepare your article for
#> submission, see the [Instructions for
#> Authors](https://journal.r-project.org/share/author-guide.pdf).
#> :::
texor::generate_rmd(your_article_path)
unlink(your_article_folder,recursive = TRUE)