latex to web

automated function for converting a single RJarticle to web

Usage

latex_to_web(
  dir,
  log_steps = TRUE,
  example = FALSE,
  auto_wrapper = TRUE,
  temp_mode = TRUE,
  web_dir = FALSE,
  interactive_mode = FALSE
)

Arguments

dir

directory path

log_steps

Enable/Disable Logging of conversion steps

example

for examples only by default keep it FALSE.

auto_wrapper

automatically creates a wrapper if TRUE, else asks user. default value TRUE

temp_mode

temp mode will convert the document in a temporary folder and keep the original article untouched. default value = TRUE

web_dir

option to create a new web directory, default FALSE

interactive_mode

interactive mode for converting articles with options. default FALSE

Value

RJweb article document in /web folder

Note

Use pandoc version greater than or equal to 2.17

Do not use example = TRUE param when working with conversions.

Examples

# Checking for pandoc version
# texor works with pandoc version >= 2.17
article_dir <- system.file("examples/article",
                 package = "texor")
dir.create(your_article_folder <- file.path(tempdir(), "tempdir"))
x <- file.copy(from = article_dir, to = your_article_folder,recursive = TRUE,)
your_article_path <- paste(your_article_folder,"article",sep="/")
texor::latex_to_web(your_article_path,log_steps = FALSE, example = TRUE, temp_mode =FALSE)
#> /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpgNFQE7/tempdir/article
#> Warning: pandoc version too old, current-v : 2.19.2 required-v : >=2.17
#> [1] FALSE
unlink(your_article_folder, recursive = TRUE)