Modify Markdown from Sweave to R-markdown

generate rmarkdown file in output folder

Usage

rnw_generate_rmd(
  article_dir,
  web_dir = TRUE,
  interactive_mode = FALSE,
  front_matter_type = "vignettes",
  autonumber_sec = TRUE,
  algorithm_render = "offline"
)

Arguments

article_dir

path to the directory which contains tex article

web_dir

option to create a new web directory, default TRUE

interactive_mode

interactive mode for converting articles with options. default FALSE

front_matter_type

knit output type for the RMarkdown file, default is "vignettes", optional for "biocstyle", "litedown"

autonumber_sec

whether to autonumber the sections, default is TRUE

algorithm_render

how to render algorithm environment with pseudocode.js, default is "offline" optional for "latest", "offline". Exclude it by using any other value

Value

R-markdown file in the web folder

Note

Use pandoc version greater than or equal to 3.1

Examples

# Note This is a minimal example to execute this function
# Please refer to texor::rnw_to_rmd for a detailed example