Sweave to RMarkdown

automated function for converting a single Sweave file to R Markdown file

Usage

rnw_to_rmd(input_file, front_matter_type = "vignettes", clean_up = TRUE)

Arguments

input_file

input Sweave file path

front_matter_type

knit output type for the RMarkdown file, default is "vignettes"

clean_up

whether to clean up the intermediate files, default is TRUE

Value

True if R Markdown file successfully generated in the same folder

Note

Use pandoc version greater than or equal to 3.1

Examples

article_dir <- system.file("examples/sweave_article", package = "texor")
dir.create(your_article_folder <- file.path(tempdir(), "tempdir"))
x <- file.copy(from = article_dir, to = your_article_folder,recursive = TRUE)
your_article_path <- paste(your_article_folder,"sweave_article", "example.Rnw",sep="/")
your_article_path <- xfun::normalize_path(your_article_path)
rnw_to_rmd(your_article_path, front_matter_type = "vignettes", clean_up = TRUE)
#> removing the unnecessary option fig=TRUE:
#>     * image,fig=TRUE, fig.cap="matrix package's image"
#> removing the unnecessary option fig=TRUE:
#>     * keep.source=TRUE, fig=TRUE
#> changing keep.source=TRUE to tidy=FALSE:
#>     * keep.source=TRUE,
#> changing \SweaveOpts{} to opts_chunk$set():
#>     * \SweaveOpts{keep.source=TRUE, fig=TRUE}
#> /usr/local/bin/pandoc +RTS -K512m -RTS /private/var/folders/hn/5bx1f4_d4ds5vhwhkxc7vdcr0000gn/T/RtmpWrderu/tempdir/sweave_article/example-knitr.Rnw --to markdown-simple_tables-pipe_tables-fenced_code_attributes --from latex --output /private/var/folders/hn/5bx1f4_d4ds5vhwhkxc7vdcr0000gn/T/RtmpWrderu/tempdir/sweave_article/example-knitr-part1.md --resource-path /private/var/folders/hn/5bx1f4_d4ds5vhwhkxc7vdcr0000gn/T/RtmpWrderu/tempdir/sweave_article -f /Users/runner/work/_temp/Library/texor/sweave_code_reader.lua --lua-filter /Users/runner/work/_temp/Library/texor/r_code_chunk_patcher.lua 
#> /usr/local/bin/pandoc +RTS -K512m -RTS /private/var/folders/hn/5bx1f4_d4ds5vhwhkxc7vdcr0000gn/T/RtmpWrderu/tempdir/sweave_article/example-knitr.Rnw --to latex --from latex --output /private/var/folders/hn/5bx1f4_d4ds5vhwhkxc7vdcr0000gn/T/RtmpWrderu/tempdir/sweave_article/example-knitr-part2.tex --resource-path /private/var/folders/hn/5bx1f4_d4ds5vhwhkxc7vdcr0000gn/T/RtmpWrderu/tempdir/sweave_article -f /Users/runner/work/_temp/Library/texor/sweave_code_remove.lua 
#> Cannot aggregate bibliography as there is no embedded bibliography
#> Found 0 Matches for target :  example
#> Found 0 Matches for target :  example
#> Found 0 Matches for target :  example\*
#> Found 0 Matches for target :  example\*
#> Found 0 Matches for target :  Sin
#> Found 0 Matches for target :  Sin
#> Found 0 Matches for target :  Sout
#> Found 0 Matches for target :  Sout
#> Found 0 Matches for target :  Scode
#> Found 0 Matches for target :  Scode
#> Found 0 Matches for target :  Sinput
#> Found 0 Matches for target :  Sinput
#> Found 0 Matches for target :  Soutput
#> Found 0 Matches for target :  Soutput
#> Found 0 Matches for target :  smallverbatim
#> Found 0 Matches for target :  smallverbatim
#> Found 0 Matches for target :  boxedverbatim
#> Found 0 Matches for target :  boxedverbatim
#> Found 0 Matches for target :  smallexample
#> Found 0 Matches for target :  smallexample
#> Found 0 Matches for target :  table\*
#> Changed \begin{table\*} to \begin{table}
#> Found 0 Matches for target :  table\*
#> Changed \end{table*} to \end{table}
#> Found 0 Matches for target :  multicolumn
#> changed \multicolumn to \multicolumnx
#> Found 0 Matches for target :  widetable
#> Changed \begin{widetable} to \begin{table}
#> Found 0 Matches for target :  widetable
#> Changed \end{widetable} to \end{table}
#> Found 0 Matches for target :  figure\*
#> Changed \begin{figure\*} to \begin{figure}
#> Found 0 Matches for target :  figure\*
#> Changed \end{figure\*} to \end{figure}
#> Found 0 Matches for target :  widefigure
#> Changed \begin{widefigure} to \begin{figure}
#> Found 0 Matches for target :  widefigure
#> Changed \end{widefigure} to \end{figure}
#> Found 0 Matches for target :  widefigure\*
#> Changed \begin{widefigure\*} to \begin{figure}
#> Found 0 Matches for target :  widefigure\*
#> Changed \end{widefigure\*} to \end{figure}
#> /usr/local/bin/pandoc +RTS -K512m -RTS /private/var/folders/hn/5bx1f4_d4ds5vhwhkxc7vdcr0000gn/T/RtmpWrderu/tempdir/sweave_article/RJwrapper.tex --to native --from latex --output /private/var/folders/hn/5bx1f4_d4ds5vhwhkxc7vdcr0000gn/T/RtmpWrderu/tempdir/sweave_article/temp-native.txt -s --resource-path /private/var/folders/hn/5bx1f4_d4ds5vhwhkxc7vdcr0000gn/T/RtmpWrderu/tempdir/sweave_article/RJwrapper.tex --lua-filter /Users/runner/work/_temp/Library/texor/find_pdf_files.lua 
#> Found 0 Matches for target :  figure\*
#> Changed \begin{figure\*} to \begin{figure}
#> Found 0 Matches for target :  figure\*
#> Changed \end{figure\*} to \end{figure}
#> Found 0 Matches for target :  widefigure
#> Changed \begin{widefigure} to \begin{figure}
#> Found 0 Matches for target :  widefigure
#> Changed \end{widefigure} to \end{figure}
#> Found 0 Matches for target :  widefigure\*
#> Changed \begin{widefigure\*} to \begin{figure}
#> Found 0 Matches for target :  widefigure\*
#> Changed \end{widefigure\*} to \end{figure}
#> Found 0 Matches for target :  algorithm
#> Changed \begin{algorithm} to \begin{figure}
#> Found 0 Matches for target :  algorithm
#> Changed \end{algorithm} to \end{figure}
#> /usr/local/bin/pandoc +RTS -K512m -RTS /private/var/folders/hn/5bx1f4_d4ds5vhwhkxc7vdcr0000gn/T/RtmpWrderu/tempdir/sweave_article/RJwrapper.tex --to markdown-simple_tables-pipe_tables-fenced_code_attributes --from latex --output /private/var/folders/hn/5bx1f4_d4ds5vhwhkxc7vdcr0000gn/T/RtmpWrderu/tempdir/sweave_article/RJwrapper.md -s --resource-path /private/var/folders/hn/5bx1f4_d4ds5vhwhkxc7vdcr0000gn/T/RtmpWrderu/tempdir/sweave_article --lua-filter /Users/runner/work/_temp/Library/texor/issue_checker.lua --lua-filter /Users/runner/work/_temp/Library/texor/abs_filter.lua --lua-filter /Users/runner/work/_temp/Library/texor/bib_filter.lua --lua-filter /Users/runner/work/_temp/Library/texor/equation_filter.lua --lua-filter /Users/runner/work/_temp/Library/texor/image_filter.lua --lua-filter /Users/runner/work/_temp/Library/texor/image_caption.lua --lua-filter /Users/runner/work/_temp/Library/texor/widetable_patcher.lua --lua-filter /Users/runner/work/_temp/Library/texor/R_code.lua --lua-filter /Users/runner/work/_temp/Library/texor/table_caption.lua --lua-filter /Users/runner/work/_temp/Library/texor/conversion_compat_check.lua --lua-filter /Users/runner/work/_temp/Library/texor/bookdown_ref.lua --citeproc 
#> Detected the following packages from article:
#>   CRAN: 
#>   Bioconductor: 
#> [1] TRUE
unlink(your_article_folder, recursive = TRUE)